Test para funcionalidad Coach Imagen

+++ SIIR/Areas/Admin/Controllers/CoachController.cs
@@ -29,58 +29,13 @@ namespace SIIR.Areas.Admin.Controllers
         }
 
         [HttpGet]
-        public IActionResult Create()
+        public IActionResult Edit(int id)
         {
-            return View();
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public IActionResult Create(Models.Coach coach)
-        {
-            //Hacer que el modelo no sea valido? Esto creo que no funciona, al menos para mi XD
-            if (ModelState.IsValid)
-            {
-                string webRootPath = _hostingEnvironment.WebRootPath;
-                var files = HttpContext.Request.Form.Files;
-                if (coach.Id == 0 && files.Count() > 0)
-                {
-                    string fileName = Guid.NewGuid().ToString();
-                    var uploads = Path.Combine(webRootPath, @"images\coaches");
-                    var extension = Path.GetExtension(files[0].FileName);
-                    
-                    using (var fileStreams = new FileStream(Path.Combine(uploads, fileName + extension), FileMode.Create))
-                    {
-                        files[0].CopyTo(fileStreams);
-                    }
-                    coach.ImageUrl = @"\images\coaches\" + fileName + extension;
-                    
-                }
-                else
-                {
-                    ModelState.AddModelError("Imagen", "Debes seleccionar una imagen");
-                }
-                _contenedorTrabajo.Coach.Add(coach);
-                _contenedorTrabajo.Save();
-                return RedirectToAction(nameof(Index));
-            }
-            return View(coach);
-        }
-
-        [HttpGet]
-        public IActionResult Edit(int? id)
-        {
-            Models.Coach coach = new Models.Coach();
-            coach = _contenedorTrabajo.Coach.GetById(id);
+            var coach = _contenedorTrabajo.Coach.GetById(id);
             
             if (coach == null)
             {
                 return NotFound();
-            }
-
-            if (id != null)
-            {
-                coach = _contenedorTrabajo.Coach.GetById(id.GetValueOrDefault());
             }
 
             return View(coach);
@@ -106,14 +61,13 @@ namespace SIIR.Areas.Admin.Controllers
 
                     var extension_new = Path.GetExtension(files[0].FileName);
 
-                    //Creamos la ruta de la imagen, quitamos la contrabarra
-                    var imagePath = Path.Combine(webRootPath, coachFromDb.ImageUrl.TrimStart('\\'));
-
-                    //Si el archivo existe
-                    if (System.IO.File.Exists(imagePath))
+                    if (!string.IsNullOrEmpty(coachFromDb.ImageUrl))
                     {
-                        //Se borra y se reemplaza con la nueva
-                        System.IO.File.Delete(imagePath);
+                        var imagePath = Path.Combine(webRootPath, coachFromDb.ImageUrl.TrimStart('\\'));
+                        if (System.IO.File.Exists(imagePath))
+                        {
+                            System.IO.File.Delete(imagePath);
+                        }
                     }
 
                     // Subimos la nueva imagen
@@ -123,9 +77,6 @@ namespace SIIR.Areas.Admin.Controllers
                     }
 
                     coach.ImageUrl = @"\images\coaches\" + fileName + extension_new;
-                    _contenedorTrabajo.Coach.Update(coach);
-                    _contenedorTrabajo.Save();
-                    return RedirectToAction(nameof(Index));
                 }
                 else
                 {
